+++ .kiro/specs/python-mud-engine/tasks.md
@@ -170,3 +170,11 @@
   - [ ] 28.1 Docker 컨테이너화
   - [ ] 28.2 설정 파일 외부화 (포트, DB 경로 등)
   - [ ] 28.3 운영 문서 작성
+
+- [ ] 29. 파일 분리 리팩토링 (한 파일 한 클래스 원칙)
+  - [x] 29.1 admin_commands.py → 클래스별 파일 분리 (14 클래스)
+  - [ ] 29.2 combat_commands.py → 클래스별 파일 분리 (5 클래스)
+  - [ ] 29.3 interaction_commands.py → 클래스별 파일 분리 (4 클래스)
+  - [ ] 29.4 npc_commands.py → 클래스별 파일 분리 (3 클래스)
+  - [ ] 29.5 combat.py → CombatManager 분리 (enum/dataclass는 유지)
+  - [ ] 29.6 repositories.py → 리포지토리별 파일 분리 (5 클래스)
