+++ src/mud_engine/commands/admin/list_templates_command.py
@@ -1,0 +1,86 @@
+# -*- coding: utf-8 -*-
+"""템플릿 목록 조회 명령어"""
+
+import logging
+from typing import List
+
+from .base import AdminCommand
+from ..base import CommandResult, CommandResultType
+from ..utils import get_user_locale
+from ...core.types import SessionType
+from ...core.localization import get_localization_manager
+
+logger = logging.getLogger(__name__)
+I18N = get_localization_manager()
+
+
+class ListTemplatesCommand(AdminCommand):
+    """템플릿 목록 조회 명령어"""
+
+    def __init__(self):
+        super().__init__(
+            name="templates",
+            description="사용 가능한 몬스터 템플릿 목록을 표시합니다",
+            aliases=["listtemplates", "tmpl"]
+        )
+
+    async def execute_admin(self, session: SessionType, args: List[str]) -> CommandResult:
+        """템플릿 목록 표시"""
+        locale = get_user_locale(session)
+        try:
+            game_engine = session.game_engine
+            if not game_engine:
+                return CommandResult(
+                    result_type=CommandResultType.ERROR,
+                    message=I18N.get_message("admin.no_game_engine", locale)
+                )
+
+            template_loader = game_engine.world_manager._monster_manager._template_loader
+            templates = template_loader.get_all_monster_templates()
+
+            if not templates:
+                return CommandResult(
+                    result_type=CommandResultType.INFO,
+                    message=I18N.get_message("admin.templates.empty", locale)
+                )
+
+            template_list = "📋 Monster Templates:\n\n"
+
+            for template_id, template_data in templates.items():
+                name_ko = template_data.get('name', {}).get('ko', '이름 없음')
+                name_en = template_data.get('name', {}).get('en', 'No name')
+                monster_type = template_data.get('monster_type', 'UNKNOWN')
+                level = template_data.get('stats', {}).get('level', 1)
+
+                template_list += f"• {template_id}\n"
+                template_list += f"  {name_ko} ({name_en})\n"
+                template_list += f"  type: {monster_type}, level: {level}\n\n"
+
+            template_list += f"Total: {len(templates)} templates\n"
+            template_list += "\nUsage: `spawnmonster <template_id> [room_id]`"
+
+            return CommandResult(
+                result_type=CommandResultType.SUCCESS,
+                message=template_list
+            )
+
+        except Exception as e:
+            logger.error(f"템플릿 목록 조회 실패: {e}")
+            return CommandResult(
+                result_type=CommandResultType.ERROR,
+                message=I18N.get_message("admin.templates.error", locale, error=str(e))
+            )
+
+    def get_help(self) -> str:
+        return """
+📋 **템플릿 목록 도움말**
+
+현재 로드된 몬스터 템플릿 목록을 표시합니다.
+
+**사용법:** `templates`
+
+**별칭:** `listtemplates`, `tmpl`
+**권한:** 관리자 전용
+
+각 템플릿의 ID, 이름, 타입, 레벨 정보를 확인할 수 있습니다.
+        """
